prototype up and running

+++ backend/src/main/java/com/nothingatall544/sunface/model/iSunFaceModel.java
@@ -1,11 +1,10 @@
 package com.nothingatall544.sunface.model;
 
 /**
- * todo implement
  * Created by derp on 7/21/2015.
  */
 public interface iSunFaceModel {
     public void syncTime();
     public SunState getSunState();
-    public double getFillPercent();
+    public float getFillPercent();
 }
